feat: includes NIS calculation
This commit improves the existing implementation by adding the
Normalized Innovation Squared (NIS) for each sensor after every
measurement timestep (k).

These values are printed to .csv files on the root folder of the app
allowing for easy import to other systems to plot it.

This commit also cleans up the code, removing unnecessary prints to
the screen that were used for debugging.

Points that still need to be addressed:
- Fine tune std_a and std_yawdd
- Write custom functions explanation on header files